% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CodesBigData.R
\docType{package}
\name{CodesBigData}
\alias{CodesBigData}
\alias{CodesBigData-package}
\title{ZERAH}
\description{
Contains BigData homework functions
}
\author{
: Yoël Zerah \email{yoel.zerah@student.isae-supaero.fr}
}
