% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZERAH.R
\name{pCertificatBigData}
\alias{pCertificatBigData}
\title{Computes the Empirical CDF from samples}
\usage{
pCertificatBigData(X)
}
\arguments{
\item{X}{Numeric vector whose elements are observations of a random variable}
}
\value{
Function representing the empirical CDF, to be evaluated in various points
}
\description{
Computes the Empirical CDF from samples
}
\examples{
n<-100 # sample number
X<-rnorm(n) # n observation of a normal law
Y<-pCertificatBigData(X) # computing ECDF
Y(0.0) # evaluating the ECDF at 0.0

}
