% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZERAH.R
\name{rCertificatBigData}
\alias{rCertificatBigData}
\title{Computes quantiles from samples}
\usage{
rCertificatBigData(n = 1000, var_mode = "Y", law = "U")
}
\arguments{
\item{n}{number of samples to return}

\item{var_mode}{"Y" to compute the sum of Random Variables, "21" for the product, "Z2" for the quotient}

\item{law}{"U" for uniform distribution, "G" (or anything else) for normal ditribution}
}
\value{
Numeric vector
}
\description{
Computes quantiles from samples
}
\examples{
n<-100 # sample number
Y<-rCertificatBigData(n,var_mode="Z1",law="G") # generating samples

}
